moved inputTag.js to the AddProjectPage folder because that is the only place it is referenced
